solhint fixes

+++ contracts/TacoCoin.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.25;
 
 import {ERC20} from "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 import {IERC20Permit} from "@openzeppelin/contracts/token/ERC20/extensions/IERC20Permit.sol";
-import "@openzeppelin/contracts/access/AccessControl.sol";
+import {AccessControl} from "@openzeppelin/contracts/access/AccessControl.sol";
 
 /// @title TacoCoin - ERC20 token with EIP-2612 permit and role-based mint/burn access
 /// @author Vladimir Gorenkov
@@ -68,8 +68,8 @@ contract TacoCoin is ERC20, IERC20Permit, AccessControl {
     function permit(
         address owner,
         address spender,
-        uint value,
-        uint deadline,
+        uint256 value,
+        uint256 deadline,
         uint8 v,
         bytes32 r,
         bytes32 s
